refactor(logic): split locked_province_assigner into part-of concern files (Refs #3290)
Phase 0 prerequisite for the colonizethis_logic package split (#3290): bring
the 602-line setup/locked_province_assigner.dart down by extracting two
concern-scoped part files, mirroring the established part-of pattern already
used by game_setup_ownership/capital_choice/game_setup_helpers.

Behavior-preserving, lexical-only extraction (no algorithm or public API
change):
- locked_province_assigner.dart (root): public surface only —
  kDefaultBacktrackLimitPerFaction, kMaxBacktracksPerLandmassBeforeCapitalRestart,
  LockedAssignerObservation, islandResidualsFeasibleGreedy,
  assignTerritoriesLockedOnLandmass + part directives.
- locked_province_assigner_graph.dart (part): pure P-P graph / feasibility
  helpers (island sizes, component size, reachable territory, ranked legal
  neighbours, phased-growth feasibility).
- locked_province_assigner_engine.dart (part): the _LockedAssignerEngine DFS
  search (tabu, phased seeding, per-faction backtrack, cross-unwind, capital
  restart) plus its trace/typedef/dfs-code internals.

All members keep their original (public/private) visibility within the single
library, so external importers and tests are unaffected.
